requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
websockets==12.0
python-multipart==0.0.9
aiofiles==23.2.1
opencv-python-headless==4.9.0.80
ultralytics==8.2.103
onnxruntime>=1.17.1
torch==2.2.2
torchvision==0.17.2
timm==0.9.16
paddlepaddle==2.6.2
paddleocr==2.8.1
numpy==1.26.4
Pillow==10.2.0
python-dotenv==1.0.1
qrcode[pil]==7.4.2
google-generativeai>=0.8.0
protobuf>=3.19.0,<4.0.0
pywin32==306; sys_platform == "win32"
pyttsx3==2.90; sys_platform == "win32"
